requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
pydantic-settings==2.2.1
python-multipart==0.0.9
pymupdf==1.24.3
google-generativeai==0.7.0
groq==0.9.0
anthropic==0.28.0
openai==1.30.0
alembic==1.13.1
